Changed the option type in BasicRequest::GetInfo to CURLINFO because it wasn't set to the correct parameter type
Removed "user/" from the URL in RedditAccess::RedditGEtRequest because not all endpoints start with it

Removed all of get_saved_items and replaced it with RetrieveSaved. It only retrieves the json for 100 elements at a time.

I added a download function and removed the mess that was the former download function