sbjson.hpp Wrapped sbjson functions in their own namespace, SBJSON
base.cpp/.hpp

Re-added base with a function, splitString, I thought didn't fit in a single class or any of the other ones

Imgur.cpp/.hpp

Moved ImgurAccess from being a class to a set a functions under the namespace ImgurAccess
Changed IsImage and IsAlbum to inline functions and fixed an issue within them where they would not return the correct result.
Added GetHash, to get the hash portion of an imgur link
Added IsDirect, to check if an imgur link is linking to a direct image. e.g i.imgur.com/image.jpeg
The endpoints for GetAlbum and GetImage where wry and contorted, so they are fixed and should work properly now

Reddit.cpp/.hpp

Wrapped RedditAccess in RST namespace, which stands for Reddit Save Tool

RedditObject.cpp/.hpp

Added Gallery struct to hold Imgur Album and Reddit Gallery images
Added ResolveImgurLinks, to call the Imgur API on the imgur links provided by Reddit and resolve them to get the images directly
Moved Saver from parsing the json from saved, to RedditObject having its own method to parse a post

Saver.cpp/.hpp
Wrapped Saver in RST namespace
Renamed WriteContent to WritePosts
Added FilterPosts to filter through with some type of criteria with args. It's not fully implemented
Addd GetListOp, which gets a list of strings in a vector from an option from agrs
Added DownloadImage, instead of simply downloading, it writes the image with extension to the disk with RedditObject::Write
Removed ImgurHandle and replaced it with a ClientId string. It will just pass ClientId to ImgurAccess functions instead of having a class
Heavy rearranged, ParseSaved to now call RedditObject::Read on each post on the Json and push the post through a switch statement to push it to the Saver::post vector and filter it through the flags in args
In WritePosts, I added Gallery and Album support and made DownloadImage get the extensions
Added --domain as an option to filter posts through which domain the URL they're linked to

CMakeLists.txt
Added base.cpp